logger/requirements.txt
Click==7.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dotenv==0.12.0
Werkzeug==1.0.0
